tidy up of server ts file and some abstraction
