+++ backend/src/config/index.ts
@@ -1,0 +1,2 @@
+export {default as createExpressApp} from "./expressSetup";
+export {default as establishMongoConnection} from "./db";
