+++ backend/src/server.ts
@@ -1,8 +1,5 @@
-import express from "express";
-import mongoose from "mongoose";
 import dotenv from "dotenv";
-import cookieParser from "cookie-parser";
-import userRoutes from "./routes/userRoutes";
+import { establishMongoConnection, createExpressApp } from "./config";
 
 // Environment variables
 if (process.env.NODE_ENV === "production") {
@@ -13,29 +10,10 @@ if (process.env.NODE_ENV === "production") {
   dotenv.config({ path: ".env.development" });
 }
 
-// Connect to MongoDB
-const establishMongoConnection = async () => {
-  try {
-    if (!process.env.MONGODB_URI) {
-      throw new Error("MONGODB_URI must be defined");
-    }
-    await mongoose.connect(process.env.MONGODB_URI);
-    console.log("Connected to MongoDB");
-  } catch (error) {
-    console.error("Error connecting to MongoDB: ", error);
-    process.exit(1);
-  }
-};
-
 establishMongoConnection();
 
-const app = express();
-app.use(express.json());
-app.use(cookieParser());
+const app = createExpressApp();
 const PORT = process.env.PORT || 3000;
-
-// Routes
-app.use("/accounts", userRoutes);
 
 app.listen(PORT, () => {
   console.log(`Server listening on port ${PORT}`);
